+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added a new API call, `PUT /metrics`, for configuring the metrics
+  system.
+
 ## Changed
 - Updated CVE-2019-3016 mitigation information in
   [Production Host Setup](docs/prod-host-setup.md)
 - In case of using an invalid JSON as a 'config-file' for Firecracker,
   the process will exit with return code 152.
 - Removed the `testrun.sh` wrapper.
+- Removed `metrics_fifo` field from the logger configuration.
 
 ## [0.21.0]